Use compiled Python library
Add script to download and compile Python from scratch. Currently, Linux
systems will use Python 3.6.6 and Mac OS systems 3.7. Once Python 3.7
compiles without SSL issues on Linux, the version should be bumped so
that ideally both operating system versions use the same Python.

The Python library is only compiled into a static library so that the
final binary should be as portable as possible without the need for
Python to be installed in a target environment.

Update Travis CI configuration to address the changed pre-reqs.

Change the way CFLAGS and LDFLAGS are populated in the Go source file
that uses the transpiled C file.