Fix staticcheck/gocyclo findings
Introduce PathElement functions to determine their respective function,
for example whether they point to a map, or a simple list.

Use new helper functions in grab function to make the code easier to read.

Introduce grabByPath function to be used if a Path object is already available.

Fix error message that start with an upper-case character.

Use typed objects in switch .(type) constructs to avoid casting in cases.