WIP Rework library API
Introduce updated external API with three different levels for read and write
operations to have a common style and naming:
- function that uses YAML Node and path string (all supported types)
- function that uses YAML Node and Path
- function that is based on File (formerly called InputFile)

Reduce external/exported structs and functions as much as possible that were
not used or should not be used.

Rename structs and types to have shorter cleaner names that are more in line
with the typical Go naming styles.

Restructure the Go source file names to be more clear on their purpose.

Fix minor code issues like unnecessary assignments in declarations.

Remove unused code and unify duplicate code.